Perfil Usuário

+++ App/semear/src/main/java/semear/dao/PermissaoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import semear.model.Grupo;
 import semear.model.Permissao;
 
-public interface PermissaoRepository extends CrudRepository<Permissao, Long> {
+public interface PermissaoRepository extends CrudRepository<Permissao, Integer> {
 	
 	List<Permissao> findByGruposIn(Grupo grupo);
